Implement proxymap queries
This is an unsupported way to query Postfix databases. The
request and response format strings are configurable in order to
give us some protection against changes in Postfix's internal
protocol, although the response parser isn't flexible enough
to handle every possible format string. The format is unlikely to
change anyway. Milter calls to Proxymap are still to be written.